server/requirements.txt
django
django-ninja
django-autoslug
django-model-utils
django-pagedown
django-cors-headers